Use pprof build tag for pprof debugging
